Add support for Git tags, fix anchors, refactor views
